Add a temporary workaround for the inconsistencies in the SigSci API
Behind the scenes, the SigSci API seems to rewrite "group" to
"multival" on site and corp rule conditions. This wouldn't necessary be
a problem, except when using "group" in TF, the persisted state will
instead contain "multival", and it will thus forever be reapplying
changes unnecessarily.

This works around the issue until SigSci can rectify the situation by
silently rewriting "group" to "multival" in the TF provider itself, and
adding a new validation so "group" can't even be used regardless.